menu list update
An update for the list of menu items. Selecting a menu item is now a separate button. Another button was added to later add the menu item to the order. A detailed view of the menu item will now be displayed to the right of the menu list. A history of the selected items, which can be deleated, appears to the right of the detail view. The menu is now displayed in its own view (for the customer) so that we can later add a separate view for the baristas.