implemented data transfer for orders
Placing an order will now save that order in the order list with all necessary details. Orders are created via POST request from the confirmation message component and accessed from the orders component via GET. The messages service handles the interaction between the orders and provides the order list  so that both components can access it.