added functionality to set order status
The status of the orders can now be accessed from the order list. Orders can be set as "Open", "Processing", or "Closed". Orders change their colors based on their status and they retain this color unless the status is changed again. The status is updated to the mock server data.