updated confirmation menu, menu list and main page
The main page was updated to include a toolbar. An update was also made to the menu items for better visibility. Icons were added to the menu item list, the menu confirmation message, as well as the toolbar. The toolbar sticks to the top of the screen (although the cards overlap the toolbar, needs to be fixed later).