added expansion box
Added an expansion box so that the orders in the order list all have the same default size. The barista can look at the details (i.e. every item that was ordered with its price) by clicking on the expansion field. Also adjusted colors in the order list for better legibility.